requirements.txt
thoughtspot-tml
PyYAML